requirements.txt
dash==2.17.1
dash_bootstrap_components==1.6.0
dash_mantine_components==0.12.1
numpy==2.0.1
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
scipy==1.14.0
SQLAlchemy==2.0.31
